requirements.txt
# Flask y extensiones
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9  # Para PostgreSQL en producción
alembic==1.13.0

# Security
Werkzeug==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0

# Production server
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2